메모리 관련 타입 일괄 변경 및 Result의 시간/메모리가 정확하도록 변경

+++ core/problem.go
@@ -9,7 +9,7 @@ type Problem struct {
 type Limit struct {
 	// second
 	TimeLimit   int
-	MemoryLimit int
+	MemoryLimit uint64
 }
 
 type TestCase struct {
